+++ packages/manifest/src/index.ts
@@ -13,8 +13,10 @@ export const name = 'Fill in the blank';
 // Function which inits element state (data property on the Content Element
 // entity)
 export const initState: DataInitializer = (): ElementData => ({
-  question: '',
+  embeds: {},
+  question: [],
   correct: [],
+  hint: '',
 });
 
 // Can be loaded from package.json
@@ -29,18 +31,15 @@ const ui = {
   forceFullWidth: true,
 };
 
-export const mocks = {
-  displayContexts: [{ name: 'No selection', data: {} }],
-};
-
 const manifest: ElementManifest = {
   type,
   version: '1.0',
   name,
   ssr: false,
+  isComposite: true,
+  isQuestion: true,
   initState,
   ui,
-  mocks,
 };
 
 export default manifest;