+++ packages/server/src/index.ts
@@ -1,5 +1,5 @@
 import type { HookServices, ServerRuntime } from '@tailor-cms/cek-common';
-import { initState, mocks, type } from '@tailor-cms/ce-fill-blank-manifest';
+import { initState, type } from '@tailor-cms/ce-fill-blank-manifest';
 import type { Element } from '@tailor-cms/ce-fill-blank-manifest';
 import every from 'lodash/every.js';
 import omit from 'lodash/omit.js';
@@ -11,12 +11,10 @@ const USER_STATE: any = {};
 
 /* eslint-disable @typescript-eslint/no-unused-vars */
 export function beforeSave(element: Element, services: HookServices) {
-  console.log('Before save hook');
   return element;
 }
 
 export function afterSave(element: Element, services: HookServices) {
-  console.log('After save hook');
   return element;
 }
 
@@ -25,7 +23,6 @@ export function afterLoaded(
   services: HookServices,
   runtime: ServerRuntime,
 ) {
-  console.log('After loaded hook');
   if (runtime === 'delivery') {
     const data = omit(element.data, ['correct']);
     return Object.assign(element, { data });
@@ -38,14 +35,12 @@ export function afterRetrieve(
   services: HookServices,
   runtime: ServerRuntime,
 ) {
-  console.log('After retrieve hook');
   return element;
 }
 
 export function beforeDisplay(element: Element, context: any) {
-  console.log('beforeDisplay hook');
-  console.log('beforeDisplay context', context);
-  return { ...context, ...USER_STATE, correct: element.data.correct };
+  if (IS_CEK) USER_STATE.correct = element.data.correct;
+  return { ...context, ...USER_STATE };
 }
 
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
@@ -54,16 +49,17 @@ export function onUserInteraction(
   context: any,
   payload: any,
 ): any {
-  console.log('onUserInteraction', context, payload);
+  const isGradable = element.data.isGradable;
   const isCorrect = every(element.data.correct, (it, i) => {
     const correct = it.map((it: string) => it.toLowerCase());
-    console.log('correct', correct, payload.response[i].toLowerCase());
     return correct.includes(payload.response[i].toLowerCase());
   });
   // Simulate user state update within CEK
   if (IS_CEK) {
     // Can be reset to initial / mocked state via UI
-    Object.assign(context, { response: payload.response, isCorrect });
+    context.response = payload.response;
+    if (isGradable) context.isCorrect = isCorrect;
+    context.isSubmitted = true;
   }
   // Can have arbitrary return value (interpreted by target system)
   // FE is updated if updateDisplayState is true
@@ -91,7 +87,6 @@ export default {
   afterRetrieve,
   onUserInteraction,
   beforeDisplay,
-  mocks,
 };
 
-export { type, initState, mocks };
+export { type, initState };